requirements.txt
numpy
pandas
streamlit==1.55.0
tensorflow
scikit-learn
matplotlib
seaborn
Pillow
joblib
